style: refactor code
