+++ frontend/src/components/recommendation/DroppableDateList.tsx
