+++ frontend/src/components/recommendation/RecommendationApp.tsx
@@ -3,7 +3,7 @@
 import Prompt from '@/components/prompt/Prompt';
 import { Box, Container, Typography } from '@mui/material';
 import createTranslation from 'next-translate/useTranslation';
-import ResultScreen from '@/components/result/ui/ResultScreen';
+import RecommendationResult from '@/components/recommendation/RecommendationResult';
 import { useEffect, useState } from 'react';
 import type { Recommendation } from '@/types/recommendation';
 import type { ApiContext } from '@/client/ApiContext';
@@ -15,7 +15,7 @@ const TRANSITION_STATE = {
   RESULT: 1,
 };
 
-const MainApp = () => {
+const RecommendationApp = () => {
   const { t } = createTranslation('common');
 
   const [recommendations, setRecommendations] = useState<Recommendation[]>([]);
@@ -68,10 +68,13 @@ const MainApp = () => {
         {step === TRANSITION_STATE.RESULT && (
           <ResultScreen recommendations={recommendations} setRecommendations={setRecommendations} />
         )} */}
-        <ResultScreen recommendations={recommendations} setRecommendations={setRecommendations} />
+        <RecommendationResult
+          recommendations={recommendations}
+          setRecommendations={setRecommendations}
+        />
       </Box>
     </Container>
   );
 };
 
-export default MainApp;
+export default RecommendationApp;
