+++ frontend/src/components/recommendation/RecommendationResult.tsx
@@ -1,18 +1,22 @@
 'use client';
 
-import React, { useState } from 'react';
-import LocationListings from './LocationListings';
+import RecommendationContainer from './RecommendationContainer';
 import Map from '@/components/map/Map';
 import { Box, Typography } from '@mui/material';
 import createTranslation from 'next-translate/useTranslation';
 import type { Recommendation } from '@/types/recommendation';
+import type { Dispatch, FC, SetStateAction } from 'react';
+import { useState } from 'react';
 
-interface ResultScreenProps {
+interface RecommendationResultProps {
   recommendations: Recommendation[];
-  setRecommendations: React.Dispatch<React.SetStateAction<Recommendation[]>>;
+  setRecommendations: Dispatch<SetStateAction<Recommendation[]>>;
 }
 
-const ResultScreen: React.FC<ResultScreenProps> = ({ recommendations, setRecommendations }) => {
+const RecommendationResult: FC<RecommendationResultProps> = ({
+  recommendations,
+  setRecommendations,
+}) => {
   const [activeStep, setActiveStep] = useState<number>(0);
   const [activeDate, setActiveDate] = useState<string>('');
   const [mapCenter, setMapCenter] = useState<{ lat: number; lng: number }>({ lat: 0, lng: 0 });
@@ -26,8 +30,8 @@ const ResultScreen: React.FC<ResultScreenProps> = ({ recommendations, setRecomme
       </Typography>
       <div style={{ height: '80vh', width: '80vw' }}>
         <div style={{ display: 'grid', gridTemplateColumns: '4fr 8fr' }}>
-          {/* Left side: LocationListings */}
-          <LocationListings
+          {/* Left side: Recommendation container */}
+          <RecommendationContainer
             recommendations={recommendations}
             setRecommendations={setRecommendations}
             activeStep={activeStep}
@@ -45,4 +49,4 @@ const ResultScreen: React.FC<ResultScreenProps> = ({ recommendations, setRecomme
   );
 };
 
-export default ResultScreen;
+export default RecommendationResult;
